move functions to Helper class to avoid circular dependency. lazy initialize remote_metadata for each Download button.
